:globe_with_meridians: Use locale codes by default
This allows us to either pass the complete locale or the language
code to configure react-intl.

Passing the locale has as a benefit that that date formatting will
be correct for the users' locale. For example 'en-GB' will use
British formatting but simply using 'en' will use American formats.

We need to make sure though that when we pass locales that oneSky
is configured correctly so that the complete locale code will still
point to the correct translation file.